Move cta hiding to woo landing page
